Turn simulation seems to work now

+++ src/game.rs
@@ -1,12 +1,12 @@
 use board::Board;
 use hero::Hero;
 
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize, Debug, Eq, PartialEq)]
 pub struct Game {
     pub id: String,
-    pub turn: u16,
+    pub turn: usize,
 	#[serde(rename="maxTurns")]
-    pub max_turns: u16,
+    pub max_turns: usize,
     pub heroes: [Hero; 4],
     pub board: Board,
     pub finished: bool,
